#0: More dep cleanup
